Assessment Generator: Reports and Security Fix (#471)
* docs: Generate assessments and apply security fix


* fix(ci): Resolve MyPy no-any-return and unused-ignore errors


* [AutoFix] Iteration 1: Apply fixes ( black)

Automated fixes applied directly to PR #471 branch.
Iteration: 1 of 10
Tools used: black

This commit was created by Jules PR AutoFix workflow.
CI verification loop will check if more fixes are needed.


* style: Reformat tests/test_debug_utils.py to fix CI failure


* fix: add mypy type ignore comments for Flask app and pytest fixtures

* fix: skip async tests when pytest-asyncio not installed

---------

+++ src/python/src/utils/debug_utils.py
@@ -326,15 +326,15 @@ def get_memory_usage() -> float:
         usage = resource.getrusage(resource.RUSAGE_SELF)
         # maxrss is in KB on Linux, bytes on macOS
         if sys.platform == "darwin":
-            return usage.ru_maxrss / (1024 * 1024)
-        return usage.ru_maxrss / 1024
+            return float(usage.ru_maxrss / (1024 * 1024))
+        return float(usage.ru_maxrss / 1024)
     except ImportError:
         # Windows fallback
         try:
             import psutil
 
             process = psutil.Process()
-            return process.memory_info().rss / (1024 * 1024)
+            return float(process.memory_info().rss / (1024 * 1024))
         except ImportError:
             return 0.0
